\name{GraphMM-package}
\alias{GraphMM-package}
\alias{GraphMM}
\docType{package}
\title{
  Graph-based mixture model for multiple hypothesis testing
}
\description{
This package performs multiple hypothesis testing for equal means using Graph-based mixture model, which takes advantage of clustering pattern to improve testing power. }
\author{
Tien Vo <tnvo@wisc.edu>, 
Michael Newton <newton@biostat.wisc.edu>
}
\references{
Link to paper
}
\keyword{ Bayes, Bayesian, empirical Bayes, graph, mixture model, hypothesis testing}
\seealso{
  Optional links to other man pages
}
